+++ src/routes/overview/OverviewHeader.styles.ts
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export const styles = {
+  headerContent: {
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
+  headerContentRight: {
+    display: 'flex',
+    justifyContent: 'end',
+  },
+} as { [className: string]: React.CSSProperties };
