--- src/routes/overview/overview.styles.ts
@@ -1,16 +1,0 @@
-// import global_BackgroundColor_light_100 from '@patternfly/react-tokens/dist/js/global_BackgroundColor_light_100';
-// import global_FontSize_md from '@patternfly/react-tokens/dist/js/global_FontSize_md';
-// import global_spacer_lg from '@patternfly/react-tokens/dist/js/global_spacer_lg';
-// import global_spacer_md from '@patternfly/react-tokens/dist/js/global_spacer_md';
-import React from 'react';
-
-export const styles = {
-  headerContent: {
-    display: 'flex',
-    justifyContent: 'space-between',
-  },
-  headerContentRight: {
-    display: 'flex',
-    justifyContent: 'end',
-  },
-} as { [className: string]: React.CSSProperties };
